requirements.txt
# Core Dependencies
numpy>=1.21.0,<1.24.0
torch>=1.12.1,<2.0.0
gymnasium>=0.26.0
stable-baselines3[extra]>=2.0.0  # [extra] includes pygame, ale-py, rich for PPO training
sb3-contrib>=2.0.0

# Simulation
# gym-pybullet-drones is installed via: pip install -e /path/to/gym-pybullet-drones
# pybullet installed via: conda install -c conda-forge pybullet (for Apple Silicon)
# mujoco>=2.3.0  # Optional, not needed for this project

# Real Drone Interface
djitellopy>=2.4.0
opencv-python>=4.7.0

# Domain Randomization
scipy>=1.9.0

# Visualization and Monitoring
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.11.0
pandas>=1.5.0

# Logging and Utilities
pyyaml>=6.0
tqdm>=4.64.0
imageio>=2.25.0
pillow>=9.3.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Documentation
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Optional: For advanced visualization
plotly>=5.13.0
moviepy>=1.0.3

# Optional: For experiment tracking
wandb>=0.13.0
mlflow>=2.1.0

# Development Tools
black>=22.12.0
flake8>=6.0.0
mypy>=0.991
